+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 🚀 Features
 - New component: AndesThumbnail | Authors: [@alejoecheguia](https://github.com/alejoecheguia)
 - Andes Message: Now allows to have a `Link Action` in the place of primary button, it has the same function as a button with a new layout defined by Andes's UX. | Authors: [@lelalesi](https://github.com/lelalesi)
+- AndesSnackbar: Fixes snackbarsQueue logic when a snackbar is no longer visible| Authors: [@EdenTorres](https://github.com/mercadolibre/fury_andesui-ios) 
 
 ## v3.5.0
 ### 🚀 Features